requirements.txt
openai>=1.30.0
aiohttp>=3.9.0
aiolimiter>=1.1.0
python-dotenv>=1.0.0
tenacity>=8.2.0
